pyproject.toml
[tool.poetry]
name = "sodabot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
jishaku = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
